検索機能を改善し、コサイン類似度の計算をドット積に変更しました。また、検索結果の表示精度を向上させ、クエリ生成メソッドを追加しました。
