+++ src/main.rs
@@ -240,13 +240,13 @@ fn main() {
         let start = Instant::now();
         let result0 = {
             let idx = index.lock().expect("Failed to lock index");
-            idx.search_cosine_similarity_parallel(&query, 16)
+            idx.search_dot_paeallel(&query, 16)
         };
         let duration = start.elapsed();
         let top_n = 20;
         println!("Search results (Top {} results, Time taken: {:.2?}):", top_n, duration);
         for (doc_id, similarity) in result0.into_iter().take(top_n) {
-            println!("Document ID: {}, Similarity: {:.4}", doc_id, similarity);
+            println!("Document ID: {}, Similarity: {:.10}", doc_id, similarity);
         }
         println!("Search results (Time taken: {:.2?}):", duration);
     }
